Charts refactored to allow update on state independently
