+++ src/App.js
@@ -1,9 +1,12 @@
 import React, {Component} from "react";
 import Grid from "@material-ui/core/Grid";
+import Paper from "@material-ui/core/Paper";
 import { makeStyles } from '@material-ui/core/styles';
 import {ChartItem} from "./ChartItem";
-import Paper from "@material-ui/core/Paper";
 import {CalendarGraphItem} from "./CalendarGraphItem";
+import {Button} from "@material-ui/core";
+import DayChart from "./DayChart";
+import WeekChart from "./WeekChart";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -19,78 +22,59 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 
-
 class App extends Component {
     constructor(props) {
-        super(props);
-        this.state = { apiResponse: [], dailyResponse: []};
+        super(props); // al react components with a constructor call super() first
+
+        this.state = {
+            fullYearProcessed: [],
+            dailyProcessed: [],
+            weeklyProcessed: [],
+            currentDate: new Date("2012-02-06T08:00:00.000Z")
+        };
     }
 
-    // callAPI() {
-    //     fetch("http://localhost:9000/api/counts")
-    //         .then(res => res.text())
-    //         .then(res => this.setState({ apiResponse: res }))
-    // }
-    fetchYearDailySums() {
-        fetch("http://localhost:9000/api/counts/year/2012")
+    componentDidMount() {
+        this.fetchYearDailySums('2012');
+    }
+
+    /* /api/counts/year/{year} */
+    fetchYearDailySums(year) {
+        fetch("http://localhost:9000/api/counts/year/" + year)
             .then(res => res.text())
             .then(res => {
                 res = JSON.parse(res);
                 let yearlyData = [[{ type: 'date', id: 'Date' }, { type: 'number', id: 'Person Count' }]]
-                let dates = res['dates'];
-                let c_in = res['in_sum'];
-                let c_out = res['out_sum'];
+                let dates = res["dates"];
+                let c_in = res["in_sum"];
+                let c_out = res["out_sum"];
 
-                for (let i = 1; i < dates.length; i++ ) {
-                    yearlyData.push([new Date(dates[i]), c_in[i]]);
+                for (let i = 0; i < dates.length; i++ ) {
+                    yearlyData.push([new Date(dates[i] + 'T08:00:00.000Z'), c_in[i]]);
                 }
-                this.setState({apiResponse: yearlyData});
+                this.setState({fullYearProcessed: yearlyData});
             });
     }
 
-    fetchSingleDayHourly() {
-        fetch("http://localhost:9000/api/counts/day/2012-05-01")
-            .then(res => res.text())
-            .then(res => {
-                res = JSON.parse(res);
-                let datasets = [];
-                console.log(res['count_in']);
-                let dialyObj = {id: 0,
-                    dataIn: res['count_in'],
-                    dataOut: res['count_out'],
-                    labels: res['labels'],
-                    title: "Hourly Counts"};
-                let weeklyObj = {
-                    id: 1,
-                    dataIn: [18094,19937,22110,25572,22937,24380,11848],
-                    dataOut: [11969,12712,14287,15629,13706,14510,8173],
-                    labels: ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'],
-                    title: "Daily Counts - One Week"
-                }
-                datasets.push(dialyObj, weeklyObj);
-
-                this.setState({dailyResponse: datasets});
-        });
-    }
-
-    componentDidMount() {
-        this.fetchYearDailySums();
-        this.fetchSingleDayHourly()
-    }
-    
+    // TODO: why are charts rendering on top of each other with each button click?
     render() {
         return (
                 <div>
-                    <h1 align={'center'}>Store Visitor Count Dashboard</h1>
-                    <Grid container spacing={2}
-                          direction={'row'}
-                          alignItems={'baseline'}>{this.state.dailyResponse.map((data) =>{
-                        return <ChartItem key={data.id} {...data}/>
-                    })}
+                    <Grid container spacing={2} direction={'row'} alignItems={'baseline'}>
+                        <Grid item xs={6}>
+                            <Paper>
+                                <DayChart/>
+                            </Paper>
+                        </Grid>
+                        <Grid item xs={6}>
+                            <Paper>
+                                <WeekChart/>
+                                {/*<Button onClick={this.onChangeDate()} variant={'outlined'}>NEXT WEEK</Button>*/}
+                            </Paper>
+                        </Grid>
                         <Grid item xs={12}>
                             <Paper>
-                                {/*<CalendarGraphItem className={"flex-col-scroll"} data={yearlyData} />*/}
-                                <CalendarGraphItem className={"flex-col-scroll"} data={this.state.apiResponse} />
+                                <CalendarGraphItem className={"flex-col-scroll"} data={this.state.fullYearProcessed} />
                             </Paper>
                         </Grid>
                     </Grid>
@@ -100,24 +84,3 @@ class App extends Component {
 }
 
 export default App;
-
-// render() {
-//     return (
-//         <div>
-//             <div>
-//                 <h1>{this.state.apiResponse}</h1>
-//             </div>
-//             <Grid container spacing={2}
-//                   direction={'row'}
-//                   alignItems={'baseline'}>{datasets.map((data) =>{
-//                 return <ChartItem key={data.id} {...data}/>
-//             })}
-//                 <Grid item xs={12}>
-//                     <Paper>
-//                         <CalendarGraphItem className={"flex-col-scroll"} data={yearlyData} />
-//                     </Paper>
-//                 </Grid>
-//             </Grid>
-//         </div>
-//     )
-// }
