+++ src/index.js
@@ -3,16 +3,25 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 import App from "./App";
-import App_Refactor from "./App_Refactor";
+
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from "@material-ui/core/Toolbar";
+import Typography from "@material-ui/core/Typography";
 
 
 ReactDOM.render(
     <React.StrictMode>
-        <App_Refactor/>
+        <AppBar position="static">
+            <Toolbar>
+                <Typography variant="h6">
+                    Visitor Count Dashboard
+                </Typography>
+            </Toolbar>
+        </AppBar>
+        <App/>
     </React.StrictMode>,
     document.getElementById('root')
 )
-
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
